+++ src/App.js
@@ -1,7 +1,7 @@
 import { BrowserRouter as Router, Switch, Route } from 'react-router-dom'
 import Home from './pages/Home'
-import Create from './pages/Create'
-import {ThemeProvider} from '@material-ui/core'
+import Search from './pages/Search'
+import {ThemeProvider, createTheme} from '@material-ui/core'
 import {grey} from '@material-ui/core/colors'
 import Layout from './components/Layout'
 import Login from './components/Login'
@@ -32,8 +32,8 @@ function App() {
             <Route exact path="/">
               <Home />
             </Route>
-            <Route path="/create">
-              <Create />
+            <Route path="/search">
+              <Search />
             </Route>
             <Route path="/login">
               <Login />
